% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{applyDevelopedPlpModel}
\alias{applyDevelopedPlpModel}
\title{Applies the exisitng plp prediction models}
\usage{
applyDevelopedPlpModel(connectionDetails, cdmDatabaseSchema,
  cohortDatabaseSchema, cohortTable, targetId, outcomeId,
  packageName = "EndometriosisRiskModels",
  modelLocation = "ccae_T6814_O6815")
}
\arguments{
\item{cdmDatabaseSchema}{The schema holding the CDM data}

\item{cohortDatabaseSchema}{The schema holding the cohort table}

\item{cohortTable}{The name of the cohort table}

\item{targetId}{The cohort definition id of the target population}

\item{outcomeId}{The cohort definition id of the outcome}

\item{packageName}{The name of the package}

\item{modelLocation}{The model location in the package}

\item{connectioDetails}{The connections details for connecting to the CDM}
}
\value{
A list with the performance and plots
}
\description{
Applies the exisitng plp prediction models
}
\details{
This will run and evaluate an exisitng plpModels
}
