% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{createCohorts}
\alias{createCohorts}
\title{Create and summarise the target and outcome cohorts}
\usage{
createCohorts(connectionDetails, cdmDatabaseSchema, cohortDatabaseSchema,
  cohortTable, targetIds = 1:2, outcomeIds = 3:4)
}
\arguments{
\item{cdmDatabaseSchema}{The schema holding the CDM data}

\item{cohortDatabaseSchema}{The schema holding the cohort table}

\item{cohortTable}{The name of the cohort table}

\item{targetIds}{The cohort definition ids of the target populations}

\item{outcomeIds}{The cohort definition ids of the outcomes}

\item{connectioDetails}{The connections details for connecting to the CDM}
}
\value{
A summary of the cohort counts
}
\description{
Create and summarise the target and outcome cohorts
}
\details{
This will create the risk prediciton cohorts and then count the table sizes
}
