% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{developEndometriosisModel}
\alias{developEndometriosisModel}
\title{Train a endometriosis model using your data}
\usage{
developEndometriosisModel(connectionDetails, cdmDatabaseSchema,
  cohortDatabaseSchema, cohortTable, cohortId, outcomeDatabaseSchema,
  outcomeTable, outcomeId, sampleSize = 5e+05, cdmVersion = 5)
}
\arguments{
\item{cdmDatabaseSchema}{The schema holding the CDM data}

\item{cohortDatabaseSchema}{The schema holding the cohort table}

\item{cohortTable}{The name of the cohort table}

\item{cohortId}{The cohort definition id of the target population}

\item{outcomeDatabaseSchema}{The schema holding the outcome table}

\item{outcomeTable}{The name of the outcome table}

\item{outcomeId}{The cohort definition id of the outcome}

\item{sampleSize}{The sample of the target population to use}

\item{cdmVersion}{The version of the cdm}

\item{connectioDetails}{The connections details for connecting to the CDM}
}
\value{
A list containing plpData and a plpResult object containing the model, prediction, evaluation summary and input settings
}
\description{
Train a endometriosis model using your data
}
\details{
This will create a new model using the same settings
}
